rename lowercase filenames of classes to camelcase like class
