+++ src/matrix/room/Room.js
@@ -1,5 +1,5 @@
 import {EventEmitter} from "../../utils/EventEmitter.js";
-import {RoomSummary} from "./summary.js";
+import {RoomSummary} from "./RoomSummary.js";
 import {SyncWriter} from "./timeline/persistence/SyncWriter.js";
 import {GapWriter} from "./timeline/persistence/GapWriter.js";
 import {Timeline} from "./timeline/Timeline.js";
