+++ src/matrix/room/RoomSummary.js
