+++ src/matrix/storage/idb/stores/MemberStore.js
